+++ src/app/admin-languages/admin-languages.component.ts
@@ -1,7 +1,8 @@
-import { Component } from '@angular/core';
+import { Component, importProvidersFrom } from '@angular/core';
 import { LanguagesService } from '../services/languages-service/languages.service';
 import { Languages } from '../models/languages/languages.model';
-import { map } from 'rxjs'; 
+import { map } from 'rxjs';
+import Swal from 'sweetalert2'; 
 @Component({
   selector: 'app-admin-languages',
   templateUrl: './admin-languages.component.html',
@@ -9,7 +10,7 @@ import { map } from 'rxjs';
 })
 export class AdminLanguagesComponent {
   itemCount : Number = 0; 
-  btntxt: string = "Agregar";
+  btntxt: string = "Add";
   goalText: string = "";
   languages: Languages[] = [];
   myLanguages: Languages = new Languages();
@@ -25,43 +26,69 @@ constructor (public languagesService: LanguagesService) {
     });
   }
 
-  AgregarJob() {
-    if (this.isEditMode && this.editId) {
-   
-      this.languagesService.updateLanguages(this.editId, this.myLanguages).then(() => {
-        console.log('Updated item successfully!');
-        this.resetForm();
-      });
-    } else {
-   
-      this.languagesService.createLanguages(this.myLanguages).then(() => {
-        console.log('Created new item successfully!');
-        this.resetForm();
-      });
-    }
-}
-
+AgregarJob() {
+          if (this.isEditMode && this.editId) {
+            // Mostrar confirmación antes de actualizar
+            Swal.fire({
+              title: 'Would you like to update this item ?',
+              text: '¿Are you sure you want to save the changes ?',
+              icon: 'warning',
+              showCancelButton: true,
+              confirmButtonText: 'Yes',
+              cancelButtonText: 'No'
+            }).then((result) => {
+              if (result.isConfirmed) {
+                this.languagesService.updateLanguages(this.editId!, this.myLanguages).then(() => {
+                  console.log('Updated item successfully!');
+                  Swal.fire('Done !', 'Updated item successfully.', 'success');
+                  this.resetForm();
+                });
+              }
+            });
+        
+          } else {
+       
+            this.languagesService.createLanguages(this.myLanguages).then(() => {
+              console.log('Created new item successfully!');
+              Swal.fire('Added !', 'The new item was added.', 'success');
+              this.resetForm();
+            });
+          }
+        }
 updateJob(id?: string) {
   const jobToEdit = this.languages.find(job => job.id === id);
   if (jobToEdit) {
     this.myLanguages = { ...jobToEdit }; 
-    this.btntxt = "Actualizar";
+    this.btntxt = "Ready for update";
     this.isEditMode = true;
     this.editId = id || null;
   }
 }
-deleteJob( id? : string) {
-this.languagesService.deleteLanguages(id).then(() => {
-  console.log('Deleted item successfully!');
-});
-console.log(id); 
-}
+
+
+deleteJob(id?: string) {
+    Swal.fire({
+      title: '¿Do you want to delete this item?',
+      text: 'This action cannot be undone.',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes',
+      cancelButtonText: 'No'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.languagesService.deleteLanguages(id).then(() => {
+          console.log('Deleted item successfully!');
+          Swal.fire('Deleted item successfully!', 'The item has been deleted.', 'success');
+        });
+      }
+    });
+  }
 
 
 
   resetForm() {
   this.myLanguages = new Languages();
-  this.btntxt = "Agregar";
+  this.btntxt = "Add";
   this.isEditMode = false;
   this.editId = null;
 }
